Let base property and base detector class able to support the case with no config.
