+++ osu.Game.Rulesets.Karaoke.Tests/Editor/Generator/BasePropertyDetectorTest.cs
@@ -8,10 +8,24 @@ using osu.Game.Rulesets.Karaoke.Edit.Generator;
 namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator;
 
 public abstract class BasePropertyDetectorTest<TDetector, TItem, TProperty, TConfig>
-    : BaseGeneratorTest<TConfig>
+    : BasePropertyDetectorTest<TDetector, TItem, TProperty>
     where TDetector : PropertyDetector<TItem, TProperty>
     where TConfig : IHasConfig<TConfig>, new()
 {
+    protected static TConfig GeneratorConfig(Action<TConfig>? action = null)
+    {
+        var config = new TConfig();
+        action?.Invoke(config);
+        return config;
+    }
+
+    protected static TConfig GeneratorDefaultConfig(Action<TConfig>? action = null)
+    {
+        var config = new TConfig().CreateDefaultConfig();
+        action?.Invoke(config);
+        return config;
+    }
+
     protected static TDetector GenerateDetector(TConfig config)
     {
         if (Activator.CreateInstance(typeof(TDetector), config) is not TDetector detector)
@@ -27,17 +41,21 @@ public abstract class BasePropertyDetectorTest<TDetector, TItem, TProperty, TCon
         CheckCanDetect(item, canDetect, detector);
     }
 
-    protected static void CheckCanDetect(TItem item, bool canDetect, TDetector detector)
-    {
-        bool actual = detector.CanDetect(item);
-        Assert.AreEqual(canDetect, actual);
-    }
-
     protected void CheckDetectResult(TItem item, TProperty expected, TConfig config)
     {
         var detector = GenerateDetector(config);
 
         CheckDetectResult(item, expected, detector);
+    }
+}
+
+public abstract class BasePropertyDetectorTest<TDetector, TItem, TProperty>
+    where TDetector : PropertyDetector<TItem, TProperty>
+{
+    protected static void CheckCanDetect(TItem item, bool canDetect, TDetector detector)
+    {
+        bool actual = detector.CanDetect(item);
+        Assert.AreEqual(canDetect, actual);
     }
 
     protected void CheckDetectResult(TItem item, TProperty expected, TDetector detector)
